Finished Exercise 3
I basically tried to write out what the for loop did using console logs. I
was confused when I saw the first console log to track what type I was
showing me all 5 items within the array. However, then I realized that it
wasnt just showing me all 5 items at the same time but each item after it
goes around the loop, ending on the 5th. And so I use an if statement
to check for when the condition of the loop is met and then to output
either true if it is or false if it isnt. And although it wasn't necessary
I displayed where we were currently at in the loop in terms of the symbol
throw so I can visualize how it is looping. And the finally, I ended with
a console log to show what the loop would look like after the iterator
happened again, which is useful for seeing which index the loop would try
going to next.